Successful merge between hex_TQ_2p and hex_TQ_cond functions
The new function, "hex_TQ", leads to the same results (and has similar
time requirements) than the old hex_TQ_cond function but is able to deal
both with 1-phase and 2-phase flows.
Added "newhex" control logical in INPUTS script that allows to switch
between the new function and the old one. This will be removed in future
commits and only the new function will be supported from now on.