Improved plotting strategy for Ts diagrams
The way in which the Ts diagrams are plot is simpler and more effective
than before.
The PTES_WRITE_CHARGE and PTES_WRITE_DISCHARGE scripts, which used to
call the "write_file" function to store the data for the Ts diagrams
inside "Plot_file.txt", are not used anymore. Instead, two new functions
have been created, "plot_Ts_diag.m" and "plot_T_storage.m", which do the
same as before without copying and reading data from an external file.
The two new functions have allowed the PTES_PLOT_CYCLE script to be
simplified; it is much easier to read and edit now.
The new strategy also has solved the problem of plotting fluids that
contain several streams, since now each stage is plotted independently
and no lines connect the end of one stream with the beginning of the
next one. In order to ensure this, the last stage of each stream must
finish with an 'end' type (rather than e.g. 'hex' or 'exp').
See how this is done inside RANK_DISCHARGE.