Defined new "hex_func" which consolidates the "hex_TQ" and "hex_TQA" functions
The new function employs the HX structure as an input and as an output
and is meant to be easy to incorporate as a method inside the new heat
exchanger class, similarly to what was done with the "compexp_func"
function.
Also, incorporating minor updates in "plot_hex", "set_hex_geom" and
"test_hex".