Integrated cold tanks inside RANK_DISCHARGE
This uses the new operation mode with the consolidated "hex_func" for
the heat exchangers. This function should in the near future replace all
the other calls employing the older "hex_TQ" function.

The Rankine cycle is called twice. The first time, it discharges using
the cold stores. The second time (once the cold stores are depleted), it
discharges against ambient. A new parameter "Load.options.useCold" has
been included inside the Load structure to differentiate the two modes.