Solved bug in RANK_DISCHARGE, and implemented 'multi-run' for condenser pressure
The error was due to CoolProp not giving accurate results for properties
very close to the steam saturation curve while using the 'TTSE'
interpolation method. This has been solved by changing to 'HEOS' backend
for the steam cycle. This is more accurate but has the inconvenient of
being slower.

After solving the bug, the 'multi-run' option (to run several cycles
while changing one parameter, for parametric studies) has been
implemented for the 'Ran_TbotC' parameter, which is the Rankine's
condenser temperature.