Introducing capability for several compressions/expansions and several tanks
In the previous version of the code, several instances of fluidH/fluidC
had to be created for having several streams of each (if, for instance,
there were several compressions/expansions).
Now, all streams of a single storage fluid are contained within the
fluid.state array, just as it is done for "gas" or for "steam".
So, now, e.g. fluidH(1) can contain several streams, and at the same
time the cycle can have several hot tanks, using fluidH(1), fluidH(2),
etc.
This required significant modification of the run_tanks and the
total_mdot function/methods, as well as all the different
CHARGE/DISCHARGE subroutines.