Introduced parasitic loss in RANKINE air condenser
This is done by defining an atmospheric 'air' fluid object, link to very
large 'atmospheric tanks'. This fluid is used together with 'hex_func'
to condense the steam stream. Then, the outlet air is put through
'compexp_func' with a very low pressure ratio to be compensate for the
pressure losses within the condenser (this is effectively like a fan,
but it is easier to code as a component placed AFTER the condenser than
BEFORE). A compressor named DFAN (discharge fan) is used.

(most of this capability was introduced within a couple of recent
commits but it is being 'activated' in this commit)