Prandtl number error 'mitigated'
CP5 was returning meaningless values for Cp and Pr around and inside the
saturation curve, which was making hex_func crash when operating with
the 'geom' model.
However, the 'weird values' makes sense because Cp and Pr are not well
defined inside the 2-phase region. A patch that substitutes
far-from-average values by the median value has been implemented, but a
more physically accurate solution is needed for the 2-phase region soon.