Correcting several bugs. JB cycle and JB-RANKINE working with new hx_class
A few things still need reviewing. For example, the "swap" capability
from hex_func has temporarily been lost and needs to be reimplemented to
be compatible with the sets of data stores in hx_class. (so care must be
put when calling hex_func, to specify before hand which is the "hot" and
which is the "cold" fluid)