Improvements on RP1 function and stream_update
'CP' mode includes a filter to automatically select CoolProp's HEOS
backend for low values of pressure. This is specially useful for the
Rankine cycle: To read fast properties away from the saturation curve
(e.g. with the 'TTSE' or 'BICUBIC' backends) and slower but accurate
properties when reading values close to the saturation curve.

'TAB' mode is now able to read all properties from the table of a given
material.

'IDL' mode has also been extended and consolidated in a way that is
consistent with the other two modes. This reuses expressions previously
found in the 'stream_update' function, which now calls RP1 and is more
concise than before.