First steps towards off-design Rankine cycle with the 'geom' heat exchanger model
Working on making modes 3, 4 and 5 operational within the 'geom' model.
This is a necessary step to be able to model the Rankine cycle
off-design.

This is still very much work in progress but I am making the commit so I
can Pull/Push and find any potential conflicts before going any further.