Making several compression/expansion stages operational
The feature of Nc_ch>1 or Ne_ch>1 had temporarily been lost when
implementing heat exchangers as objects. Now heat exchangers are
constructed within the INITIALISE function. The total number of heat
exchangers depends on Nc_ch and Ne_ch, as does the index allocated to
each device. This seems to be working well for the JB-PTES and
JB-RANKINE cycles in design mode. Needs further testing for the
off-design mode.