Set up global TABS structure and adapted names for some .dat files
The TABS structure contains the tabulated information on the materials
created by create_table. TABS is a global variable (a preliminary form
of a library) that can be accessed in any script/function that declares
it. Currently (and ideally), only create_table and RP1 use the TABS
structure.

This is useful because it enables the RP1 function to access the TABS
library directly, without one having to send it the tabulated data as an
input.

The 'valid_name' function is used to simplify some CoolProp names to use
as properties of the TABS structure. For example, 'INCOMP::MEG2[0.56]'
is transformed into 'INCOMP_MEG2_056'.