Updated stream_update calls to not require the 'fluid' structure
This is important for the future version of the set_hex/set_hex_geom
function. If a heat exchanger was initially run in a design mode (e.g.
'eff' or 'DT') and, after cycle convergence, one wants to obtain the
required heat exchanger geometry (for costing purposes), it is important
to be able to able to call set_hex with only the HX structure as an
input (not requiring the fluidH and fluidC structures, as this could
easily lead to human error while tracing the original hex_func call in
other scripts). The HX structure stores the names of the fluids employed
(but not the whole fluid structures), which is enough to run the new
version of stream_update.