Implementation of new set_hex_geom algorithm into PTES code
The 'set_hex' calls have been changed back to 'hex_func'. The 'hex_func'
function itself creates the heat exchanger geometry, when necessary
(i.e. when using the 'geom' model), via the 'set_hex_geom' function.

When using a heat exchanger in 'eff' or 'DT' models, the geometry is
computed only a posteriory after solving the cycle. This is achieved by
calling 'set_hex_geom' within the ECONOMICS script. This has the
advantage that the obtained geometry corresponds to the converged
solution.