Code running with 'geom' model again.
There was a problem with how 'mCmin' was defined inside hex_func in mode
5 (used for heat rejection).

The code is fairly slow because the rejection air streams require the
HEOS interface (it seems that no interpolation tables can be created in
CoolProp for the air mixture).

It is possible to get the code running faster by using 'Nitrogen'
instead of 'Air' as an approximation and use 'BICUBIC&HEOS'
interpolation, but a significant second-law error appears.