Implemented secondary HTF loop in heat pump for JB-RANKINE
The cold tanks employ 'Water' and the intermediate heat transfer fluid
(HTF) is 'INCOMP::MEG2[0.56]' (i.e. an eutectic solution of ethylene
glycol in water).

Two heat exchangers are employed at the cold side, one between the
working fluid (Nitrogen) and the HTF, and another one between the HTF
and the cold tanks (Water). The hot side of the HTF has a temperature
close to the discharged temperature of the cold tanks (typically T0).
The mass flow rate of the HTF is selected as to ensure that its minimum
temperature is just above the freezing point of water, as to avoid
freezing in the Water-HTF heat exchanger.