Preliminary integration of tabulated steam properties ('LIB' library)
The new 'LIB' folder and files contain the creation routines and Data,
which at the moment are being called from within the RP1 function.

This has been done for Water/Steam and only for two kinds of tables:
enthalpy-pressure ('HmassP_INPUTS') and pressure-quality ('PQ_INPUTS').

There is still a lot of scope for improvement, in terms of speed and
flexibility (i.e. other tables and other fluids) but the results seem
quite promising so far.