Improvements to LIB_THERM
An attempt was made to gain speed by compiling the RLIB method. However,
this didn't show any significant gain, and therefore the
'griddedInterpolant' calls were re-introduced (as they had been removed
in previous commit to be able to compile RLIB). However, they are now
created inside the LIB_THERM constructor, which saves time when calling
RLIB.

Some improvements were made on the distribution and coarseness of the
grids around the saturation dome.