Better pressure loss estimation in HEX 'eff' model
The pressure losses in the secondary channel of the heat exchanger, when
using the 'eff' model, can be estimated to be a fraction of 'ploss' (the
design pressure loss for the primary channel). This is done using a
pressure loss factor that is geometry independent (depends on the ratio
of mass flow rates, the ratio of densities and the ratio of inlet
pressures).
This results in a better estimate than assuming equal pressure loss for
the two channels. However, the estimate is an approximation, as detailed
knowledge about the geometry would be required to determine whether the
flows are laminar or turbulent and include the friction coefficients
when computing the pressure loss factor.