Introducing cross-flow heat exchangers
Most of the new code is contained inside the new 'hex_compute_Xflow'
function, which substitutes 'hex_compute_area' for cross-flow heat
exchangers, and a new section inside developed_flow. A semi-analytical
method is followed where the heat transfer and friction factor
coefficients are computed along each channel, and an analytical solution
is used to correlate NTU and effectiveness (assuming constant Cp). The
process requires iteration.

The current model works in both directions: creating a geometry for
given performance objective, and predicting performance for a given
geometry. It has been tested satisfactorily for an air-to-air and
air-to-water heat rejection unit, but still needs to be tested for
air-to-steam condenser.