Added separate generator and motor for Load.mode=3
For solar-PTES cycle, the Rankine cycle will already have a generator installed. Therefore, only a motor needs to be added to the system. Each of these devices will have been designed for their specific power ratings and will therefore operate close to the design point. As it turns out, this only has a small impact on efficiency.