Add a calculation to find electrical heat addition to hot storage
In some cases, the hot fluid is discharged to a temperature below its original temperature. Technically, this constitutes a loss as heat needs to be added back to the fluid to return it to the correct temperature. We need to add this functionality properly since this loss isn't currently calculated. In the mean time, a small addition has been made to ENERGY_BALANCE_v2 to calculate this required heat addition, and recalculate the round-trip efficiency based on this heat being provided by an electrical heater.