Calculate discharging pressure ratio automatically
Added an option to the inputs that attempts to calculate the discharging pressure ratio after charge has occured. This method avoids the hot fluid being overcooled in discharge. It also extracts most of the energy in the hot fluid. However, this method does not calculate the optimal discharge pressure ratio, so it is possible to improve efficiency further.