Add steam turbine costs for Load,mode=3
Steam turbine costs have been added to the economic calculations for solar-PTES (Load.mode=3) that is NOT a retrofit. Rather than calculating the cost of the compressors, expanders, heat exchangers in the steam turbine, a fixed steam turbine package cost is assumed (in the region of 500-1000 $/kW). A generator cost is still calculated, but could be removed in the future. The calculation of component costs for the retrofit has also been modified, as previously all heat exchanger costs were being calculated