Supercritical Rankine script and a few other changes
This commit includes a new SUPER_RANK_DISCHARGE script that was made to
see what happens when two reheating processes are used in the steam
cycle, meant to improve the efficiency when operating at supercritical
pressures. However, the improvement is only modest because the different
pressure levels for each turbine and steam extraction points seem to
fall far from optimum. Signifcant more work is probably needed to get
this working well.

Other minor changes are made in several files.