Improvements to JB_CHARGE in Load.mode==3 and ENERGY_BALANCE
This has removed the heat rejection unit for the heat pump of the
Brayton-Rankine system (JB_CHARGE in Load.mode==3).

Also, the Brayton-Rankine system can operate now without the cold stores
(but the economics are not running well yet in that scenario!).

Several changes were made to ENERGY_BALANCE to properly compute COP,
heat engine efficiency and exergetic efficiency in Load.mode==3,
depending on whether the cold stores are used or not, and if parasitics
are accounted for or not.