Add discharge pressure ratio to design iterator to help find optimal PRdis.
This is done by trying to match the charge and discharge hot fluid mass flow rate. If the temperatures are chosen correctly (via PRdis) then the charge and discharge hot fluid mass flow rates are the same ensuring that the charge and discharge durations are equal and the tanks will be fully discharged. (Plus the code already tries to ensure that the hot fluid is returned to its original temperature). Thus, the optimal PRdis is the one that helps return the cycle to its original state. Presumably, this only works at the moment for systems with equal charge and discharge durations and therefore needs to be further updated for more general cases.