Modify PRdis iteration in discharge mode.
The discharge pressure ratio is iterated over to find an optimal value. The previuos method (see earlier commit) does not work effectively if charge and discharge durations are different. A new method which aims to return the cold storage fluid to its initial temperature is used. (This is for design point calculations only).