Add method for cycle to produce the required power output during discharge.
The program works by specifying the working fluid mass flow rate. Therefore, the discharging power output is an output of the program. A method to iterate the code, updating the mass flow rate, in order to achieve a desired discharge power output has been added. This is similar to the method that was only available for the simple_interface, but now it can be used even if not using the simple interface. This version also replaces an earlier implementation where previously generated output files were read in (Lmulti_mdot).